client/requirements.txt
streamlit
rich
python-dotenv
openrouter
openai
scikit-learn
flask
flask-smorest
flask-cors
seaborn
matplotlib
pandas
numpy
imbalanced-learn
pydeck
requests